requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docker-pycreds==0.4.0
easyocr==1.7.2
einops==0.8.1
einx==0.3.0
entmax==1.3
eval_type_backport==0.2.2
filelock==3.13.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
jsonargparse==3.13.1
lazy_loader==0.4
lightning-utilities==0.14.2
lmdb==1.6.2
lpips==0.1.4
lxml==5.3.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.2.0
networkx==3.2.1
ninja==1.11.1.4
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
portalocker==3.1.1
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacrebleu==2.5.1
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.24.1
setproctitle==1.3.5
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchmetrics==0.10.3
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.50.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
unbabel-comet==2.2.4
urllib3==2.3.0
vector-quantize-pytorch==1.22.3
wandb==0.19.8
yarl==1.18.3
